Updated Format to Follow Java Standard Style Guides
https://google.github.io/styleguide/javaguide.html#s4.1.1-braces-always-used
says "Braces are used with if, else, for, do and while statements, even when the body is empty or contains only a single statement."

and so does the https://www.oracle.com/technetwork/java/javase/documentation/codeconventions-142311.html#431
which shows every statment having the bracket on the same line of the statement and then after the inside on a seperate line. Even if the code is one line long.